Split report viewer component (one basic and one with dialog) into two (#58)
